requirements.txt
discord.py
speedtest-cli
python-dotenv